LPS-77456 Move update handlers to internal.
